Style: add margin/padding variables
For now it is reasonable to assume that we will not have different sizes
for marging and paddings.
My knowledge of UI design is quite limited, so I'll not dive deep into
that.
This commit adds a fibonacci based sequence of margin sizes that should
be used for both margins and paddings throughout the system.