Purchase: nested attr for payments and products
Ideally data about products and payments should be entered in the
purchase form. This allows the user not to think about the relationship
between these and also not to have to create a product previous to
creating a purchase.

This still needs to improve a lot if we are to provide some intelligent
reports about the purchases.