Purchase: add due_at, paid_at, due_amount and paid_amount
We need to have due dates, paid dates, due amounts and paid amounts.

This may not be complex enough for the use case, but it is certainly an
improvement over past situation.

It is possible (likely) that Payment will have to be a model of its own.

It happens sometimes that two payments are made to a single debt, which
is a scenario that is not well covered here.

Given this is so far an excercise in learnin Rails it is nice to have
these mutating requirements as these are common in real life and they
surface problem I need to learn how to tackle.