Treatment: initial modelling of treatment administration
This is the first commit of the Treatment Administration feature.

The goal is to provide a means to control the administration of
treatment, specially medication, to our dogs. It is quite likely that
the same feature could be used to manage administration to ourselves, so
some attention is also given to make it expansible in the future.